+++ src/main/java/souris/botdealer/ui/Section.java
@@ -13,7 +13,7 @@ public enum Section {
 	EVENTS("events", "/fxml/sections/events.fxml"),
 	WALLETS("wallets", "/fxml/sections/wallets.fxml"),
 	MUSIC("music", "/fxml/sections/placeholder.fxml"),
-	TTS("tts", "/fxml/sections/placeholder.fxml"),
+	TTS("tts", "/fxml/sections/tts.fxml"),
 	GUILDS("guilds", "/fxml/sections/guilds.fxml"),
 	SETTINGS("settings", "/fxml/sections/settings.fxml"),
 	ABOUT("about", "/fxml/sections/about.fxml");